Reorganize signer related code
