+++ app/src/main/kotlin/net/primal/android/signer/SignerMethod.kt
@@ -1,0 +1,6 @@
+package net.primal.android.signer
+
+enum class SignerMethod(val method: String) {
+    GET_PUBLIC_KEY("get_public_key"),
+    SIGN_EVENT("sign_event"),
+}
